fix: make animation not break pitch calculations
This applies for the animations "bike.balloon.inflate" and "bike.balloon.deflate". For this fix, the original inflate animation was set to be a play once animation, a new looping animation was added, called "bike.balloon.inflate.hold". Additionally, a method for preventing pitch calculations from happening by overwrite was added.