refactor(DriveForceComponent): add DrivetrainState and remove hidden dependencies
Added the new DrivetrainState record that contains the settings for the component and grouped it with DriveForceComponent.
Updated DriveForceComponent to use the new record so that there is no hidden dependency anymore.
Classes must now add a DrivetrainState via `this.addStateComponent()` to use this force, else it returns 0f.
`DrivetrainState.default()` initializes with the same values previously defined as static constants.